+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # BOT CHANGELOG
 
+## [v2.1.0] - 2025-12-31
+- `aliasdb` and `aliastk` commands
+- Discord bot - updated the status
+- `feedback` - 20% chance donation message is included
+
 ## [v2.0.3] - 2025-12-24
 - fixed infil glossary image layout